backend/requirements.txt
# Python 3.11 Required

# Core Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.36
alembic==1.14.0
# PostgreSQL driver (install with Python 3.11 when using Postgres)
# psycopg2-binary==2.9.9

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Document Processing (non-AI)
pdfplumber==0.10.3
python-docx==0.8.11
langdetect==1.0.9

# Authentication
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# Async Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
requests==2.31.0
numpy==1.26.2

# AI/ML - Embeddings & Vector Search
openai==1.3.5                    # For embeddings using text-embedding-3-small
faiss-cpu==1.7.4                 # Vector similarity search (use faiss-gpu if CUDA available)
tiktoken==0.5.2                  # Token counting for OpenAI models
sentence-transformers==2.2.2     # Alternative: local embeddings (MiniLM-L6-v2)

# Machine Learning & NLP
scikit-learn==1.3.2              # For lightweight ML tasks, similarity metrics
nltk==3.8.1                      # Natural language toolkit for tokenization, sentence splitting

# LLM & Prompting Frameworks
langchain==0.1.0                 # LLM orchestration and chain management
langchain-openai==0.0.2          # OpenAI integration for LangChain
langchain-anthropic==0.0.1       # Anthropic Claude integration (optional)
# Note: Install additional providers as needed:
# langchain-cohere, langchain-huggingface, langchain-google-vertexai, etc.
